requirements.txt
#os
#argparse
pandas==2.2.3
matplotlib==3.10.1
numpy==2.2.4            # Numeric
tk==0.1.0
configparser==7.2.0
scipy==1.15.2
ttkthemes==3.2.2
customtkinter==5.2.2
openpyxl==3.1.5         # for Excel
pillow==11.1.0          # for Images
# pyinstaller
# altgraph              # for PyInstaller
